Fixing segfaults in GenerateJSON and compiler warnings
This fixes a segfault in GenerateJSON that appears to come from
hammering the disk too hard when opening and closing files very rapidly.
By simply recognizing that we're opening and closing the same file over
and over again, we can lift up the file manipulation and just pass
around a file handle. This solves the segfault.

At the same time, patching a couple of compilation warnings.