Use sorted with key option for broader ROOT support
Fixes: TypeError: '<' not supported between instances of 'TH1F' and 'TH1F'

Applying `sort` to a list of TH1s is broken in ROOT v6.20, so use
instead use `sorted` with a key of ROOT.TObject.GetName to sort the
histograms alphabetically.

Note: If hists is a list of histogram objects
hists.sort() will sort alphabetically by name in ROOT v6.10 AND v6.22,
but not in v6.20 and at the time of MR !70 HistFitter is unable to
compile in ROOT v6.22. `sorted` is a more broadly applicable function
and has the advnatage of not mutating objects, so is probably a better
solution in general as it gives the availability to be explicit on _how_
things are sorted.